Add root-level nodebox binary for simpler GUI invocation
- Add nodebox binary at workspace root that wraps nodebox-gui
- Convert nodebox-gui to a pure library crate with run() function
- Rename nodebox-cli binary to avoid name conflict
- Running the GUI is now just `cargo run`
