+++ src/main.rs
@@ -1,0 +1,3 @@
+fn main() -> eframe::Result<()> {
+    nodebox_gui::run()
+}
